+++ kdiff3/src-QT4/fileaccess.cpp
@@ -1870,7 +1870,7 @@ void ProgressDialog::setInformation(const QString& info, double dCurrent, bool b
       return;
    ProgressLevelData& pld = m_progressStack.back();
    pld.m_dCurrent = dCurrent;
-   int level = m_progressStack.size();
+   size_t level = m_progressStack.size();
    if ( level==1 )
    {
       m_pInformation->setText( info );
@@ -1888,7 +1888,7 @@ void ProgressDialog::setInformation(const QString& info, bool bRedrawUpdate )
    if ( m_progressStack.empty() )
       return;
    //ProgressLevelData& pld = m_progressStack.back();
-   int level = m_progressStack.size();
+   size_t level = m_progressStack.size();
    if ( level==1 )
    {
       m_pInformation->setText( info );
@@ -1989,7 +1989,7 @@ void ProgressDialog::recalc( bool bUpdate )
       killTimer( m_progressDelayTimer );
    m_progressDelayTimer = startTimer( 3000 ); /* 3 s delay */
 
-   int level = m_progressStack.size();
+   size_t level = m_progressStack.size();
    if( ( bUpdate && level==1) || m_t1.elapsed()>200 )
    {
       if (m_progressStack.empty() )
